Tried to move more validation out of the requests but I was getting async pipeline issues. Also moved namespaces around. Should have turned this into more than one commit.
