--- Clobo.Application/Agents/Commands/DeleteAgent/DeleteAgentCommandValidator.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using FluentValidation;
-
-namespace Clobo.Application.Agents.Commands.DeleteAgent
-{
-    public class DeleteAgentCommandValidator : AbstractValidator<DeleteAgentCommand>
-    {
-        public DeleteAgentCommandValidator()
-        {
-            RuleFor(x => x.Id)
-                .NotEmpty().WithMessage("Id cannot be empty")
-                .GreaterThan(0).WithMessage("Id must be greater than 0");
-        }
-    }
-}
-
