--- Clobo.Application/Teams/Commands/AddTeam/AddTeamCommandValidator.cs
@@ -1,16 +1,0 @@
-using FluentValidation;
-using FluentValidation.AspNetCore;
-using Microsoft.AspNetCore.Builder;
-
-namespace Microsoft.Extensions.DependencyInjection.Teams.AddTeam.Commands;
-
-public class AddTeamCommandHandlerValidtor : AbstractValidator<AddTeamCommand>
-{
-    public AddTeamCommandHandlerValidtor()
-    {
-        RuleFor(x => x.Name)
-            .NotEmpty().WithMessage("Name cannot be empty")
-            .MinimumLength(4).WithMessage("Name must be 4 characters long.")
-            .MaximumLength(25).WithMessage("Name cannot be more than 25 characters");
-    }
-}
