--- Clobo.Application/Teams/Commands/DeleteTeam/DeleteTeamCommandValidator.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using FluentValidation;
-
-namespace Clobo.Application.Teams.Commands.DeleteTeam
-{
-    public class DeleteTeamCommandValidator : AbstractValidator<DeleteTeamCommand>
-    {
-        public DeleteTeamCommandValidator()
-        {
-            RuleFor(x => x.Id)
-                .NotEmpty().WithMessage("Id cannot be empty")
-                .GreaterThan(0).WithMessage("Id must be greater than 0");
-        }
-    }
-}
-
