+++ Clobo/Controllers/TeamAgentController.cs
@@ -2,9 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Clobo.Application.TeamAgents.Commands.AddAgentToTeam;
-using Clobo.Application.TeamAgents.Commands.AgentTeamLead;
-using Clobo.Application.TeamAgents.Commands.RemoveAgentFromTeam;
+using Clobo.Application.Business.TeamAgents.Commands.AddAgentToTeam;
+using Clobo.Application.Business.TeamAgents.Commands.AgentTeamLead;
+using Clobo.Application.Business.TeamAgents.Commands.RemoveAgentFromTeam;
+using Clobo.Domain.Entities;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 
@@ -15,6 +16,7 @@ namespace Clobo.Controllers
     public class TeamAgentController : ApiControllerBase
     {
         [HttpPut("/addAgentToTeam")]
+        [ProducesResponseType(typeof(TeamAgent), StatusCodes.Status200OK)]
         public async Task<IActionResult> AddAgentToTeam([FromBody] AddAgentToTeamCommand command)
         {
             var res = await Mediator.Send(command);
@@ -22,6 +24,7 @@ namespace Clobo.Controllers
         }
 
         [HttpPut("/agentTeamLead")]
+        [ProducesResponseType(typeof(TeamAgent), StatusCodes.Status200OK)]
         public async Task<IActionResult> MakeAgentTeamLead([FromBody] AgentTeamLeadCommand command)
         {
             var res = await Mediator.Send(command);
@@ -29,6 +32,7 @@ namespace Clobo.Controllers
         }
 
         [HttpPut("/removeAgentFromTeam")]
+        [ProducesResponseType(typeof(bool), StatusCodes.Status200OK)]
         public async Task<IActionResult> MakeAgentTeamLead([FromBody] RemoveAgentFromTeamCommand command)
         {
             var res = await Mediator.Send(command);
