+++ Clobo/Controllers/TeamController.cs
@@ -2,14 +2,14 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Clobo.Application.Business.Teams.Commands.AddTeam;
+using Clobo.Application.Business.Teams.Commands.DeleteTeam;
+using Clobo.Application.Business.Teams.Commands.UpdateTeam;
+using Clobo.Application.Business.Teams.Requests.GetAllTeams;
 using Clobo.Application.Common.Interfaces;
-using Clobo.Application.Teams.Commands.DeleteTeam;
-using Clobo.Application.Teams.Commands.UpdateTeam;
-using Clobo.Application.Teams.GetAllTeams.Requests;
 using Clobo.Domain.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.DependencyInjection.Teams.AddTeam.Commands;
 using NuGet.Protocol.Core.Types;
 
 namespace Clobo.Controllers
